requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
peft>=0.8.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Qwen VLM specific
qwen-vl-utils>=0.0.1

# Training & Optimization
deepspeed>=0.12.0
wandb>=0.16.0

# Data Processing
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.24.0
scipy>=1.11.0

# Evaluation
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
